refactor(web): apply loaders and suspense with restructured architecture
- Convert async API functions to queryOptions objects for reusability in prefetch and query
- Add queryClient to router context to enable loader integration
- Split components into smaller units to support suspense boundaries
- Restructure routes to better support TanStack Router loaders
