+++ web/src/api/github.ts
@@ -1,126 +1,112 @@
 import type { Endpoints } from "@octokit/types";
 
+import { queryOptions } from "@tanstack/react-query";
+
 import { baseUrl } from ".";
 
-export async function readBranch(
-  owner: string,
-  repo: string,
-  branch: string,
-  signal?: AbortSignal | null,
-) {
-  const response = await fetch(
-    `${baseUrl}/api/github/repos/${owner}/${repo}/branches/${branch}`,
-    { signal },
-  );
-  if (!response.ok) {
-    throw new Error(response.statusText);
-  }
-  const data =
-    (await response.json()) as Endpoints["GET /repos/{owner}/{repo}/branches/{branch}"]["response"]["data"];
-  return data;
+export function readBranchCommits(owner: string, repo: string, branch: string) {
+  return queryOptions({
+    queryFn: async ({ signal }) => {
+      const response = await fetch(
+        `${baseUrl}/api/github/repos/${owner}/${repo}/branches/${branch}/commits`,
+        { signal },
+      );
+      if (!response.ok) {
+        throw new Error(response.statusText);
+      }
+      const data =
+        (await response.json()) as Endpoints["GET /repos/{owner}/{repo}/commits"]["response"]["data"];
+      return data;
+    },
+    queryKey: ["branch-commits", owner, repo, branch],
+  });
 }
 
-export async function readBranchCommits(
-  owner: string,
-  repo: string,
-  branch: string,
-  signal?: AbortSignal | null,
-) {
-  const response = await fetch(
-    `${baseUrl}/api/github/repos/${owner}/${repo}/branches/${branch}/commits`,
-    { signal },
-  );
-  if (!response.ok) {
-    throw new Error(response.statusText);
-  }
-  const data =
-    (await response.json()) as Endpoints["GET /repos/{owner}/{repo}/commits"]["response"]["data"];
-  return data;
+export function readBranches(owner: string, repo: string) {
+  return queryOptions({
+    queryFn: async ({ signal }) => {
+      const response = await fetch(
+        `${baseUrl}/api/github/repos/${owner}/${repo}/branches`,
+        { signal },
+      );
+      if (!response.ok) {
+        throw new Error(response.statusText);
+      }
+      const data =
+        (await response.json()) as Endpoints["GET /repos/{owner}/{repo}/branches"]["response"]["data"];
+      return data;
+    },
+    queryKey: ["branches", owner, repo],
+  });
 }
 
-export async function readBranches(
-  owner: string,
-  repo: string,
-  signal?: AbortSignal | null,
-) {
-  const response = await fetch(
-    `${baseUrl}/api/github/repos/${owner}/${repo}/branches`,
-    { signal },
-  );
-  if (!response.ok) {
-    throw new Error(response.statusText);
-  }
-  const data =
-    (await response.json()) as Endpoints["GET /repos/{owner}/{repo}/branches"]["response"]["data"];
-  return data;
+export function readCommit(owner: string, repo: string, ref: string) {
+  return queryOptions({
+    queryFn: async ({ signal }) => {
+      const response = await fetch(
+        `${baseUrl}/api/github/repos/${owner}/${repo}/commits/${ref}`,
+        { signal },
+      );
+      if (!response.ok) {
+        throw new Error(response.statusText);
+      }
+      const data =
+        (await response.json()) as Endpoints["GET /repos/{owner}/{repo}/commits/{ref}"]["response"]["data"];
+      return data;
+    },
+    queryKey: ["commit", owner, repo, ref],
+  });
 }
 
-export async function readCommit(
-  owner: string,
-  repo: string,
-  ref: string,
-  signal?: AbortSignal | null,
-) {
-  const response = await fetch(
-    `${baseUrl}/api/github/repos/${owner}/${repo}/commits/${ref}`,
-    { signal },
-  );
-  if (!response.ok) {
-    throw new Error(response.statusText);
-  }
-  const data =
-    (await response.json()) as Endpoints["GET /repos/{owner}/{repo}/commits/{ref}"]["response"]["data"];
-  return data;
+export function readPullCommits(owner: string, repo: string, prNumber: number) {
+  return queryOptions({
+    queryFn: async ({ signal }) => {
+      const response = await fetch(
+        `${baseUrl}/api/github/repos/${owner}/${repo}/pulls/${prNumber}/commits`,
+        { signal },
+      );
+      if (!response.ok) {
+        throw new Error(response.statusText);
+      }
+      const data =
+        (await response.json()) as Endpoints["GET /repos/{owner}/{repo}/pulls/{pull_number}/commits"]["response"]["data"];
+      return data;
+    },
+    queryKey: ["pull-commits", owner, repo, prNumber],
+  });
 }
 
-export async function readPullCommits(
-  owner: string,
-  repo: string,
-  pullNumber: number,
-  signal?: AbortSignal | null,
-) {
-  const response = await fetch(
-    `${baseUrl}/api/github/repos/${owner}/${repo}/pulls/${pullNumber}/commits`,
-    { signal },
-  );
-  if (!response.ok) {
-    throw new Error(response.statusText);
-  }
-  const data =
-    (await response.json()) as Endpoints["GET /repos/{owner}/{repo}/pulls/{pull_number}/commits"]["response"]["data"];
-  return data;
+export function readPulls(owner: string, repo: string) {
+  return queryOptions({
+    queryFn: async ({ signal }) => {
+      const response = await fetch(
+        `${baseUrl}/api/github/repos/${owner}/${repo}/pulls`,
+        { signal },
+      );
+      if (!response.ok) {
+        throw new Error(response.statusText);
+      }
+      const data =
+        (await response.json()) as Endpoints["GET /repos/{owner}/{repo}/pulls"]["response"]["data"];
+      return data;
+    },
+    queryKey: ["pulls", owner, repo],
+  });
 }
 
-export async function readPulls(
-  owner: string,
-  repo: string,
-  signal?: AbortSignal | null,
-) {
-  const response = await fetch(
-    `${baseUrl}/api/github/repos/${owner}/${repo}/pulls`,
-    { signal },
-  );
-  if (!response.ok) {
-    throw new Error(response.statusText);
-  }
-  const data =
-    (await response.json()) as Endpoints["GET /repos/{owner}/{repo}/pulls"]["response"]["data"];
-  return data;
-}
-
-export async function readRepository(
-  owner: string,
-  repo: string,
-  signal?: AbortSignal | null,
-) {
-  const response = await fetch(
-    `${baseUrl}/api/github/repos/${owner}/${repo}`,
-    { signal },
-  );
-  if (!response.ok) {
-    throw new Error(response.statusText);
-  }
-  const data =
-    (await response.json()) as Endpoints["GET /repos/{owner}/{repo}"]["response"]["data"];
-  return data;
+export function readRepository(owner: string, repo: string) {
+  return queryOptions({
+    queryFn: async ({ signal }) => {
+      const response = await fetch(`${baseUrl}/api/github/repos/${owner}/${repo}`, {
+        signal,
+      });
+      if (!response.ok) {
+        throw new Error(response.statusText);
+      }
+      const data =
+        (await response.json()) as Endpoints["GET /repos/{owner}/{repo}"]["response"]["data"];
+      return data;
+    },
+    queryKey: ["repository", owner, repo],
+  });
 }
