+++ web/src/main.tsx
@@ -12,7 +12,7 @@ dayjs.extend(relativeTime);
 const queryClient = new QueryClient();
 
 // Create a new router instance
-const router = createRouter({ routeTree });
+const router = createRouter({ context: { queryClient }, routeTree });
 
 // Register the router instance for type safety
 declare module "@tanstack/react-router" {
